Use `int64_t` in place of `int32_t` for line numbers.

+++ asteria/src/compiler/parser_error.hpp
@@ -11,7 +11,7 @@ namespace Asteria {
 class Parser_Error
   {
   public:
-    enum Code : std::uint32_t
+    enum Code : std::int64_t
       {
         // Special
         code_success                                    =    0,
@@ -74,19 +74,19 @@ class Parser_Error
     ROCKET_PURE_FUNCTION static const char* describe_code(Code xcode) noexcept;
 
   private:
-    std::uint32_t m_line;
+    std::int64_t m_line;
     std::size_t m_offset;
     std::size_t m_length;
     Code m_code;
 
   public:
-    constexpr Parser_Error(std::uint32_t xline, std::size_t xoffset, std::size_t xlength, Code xcode) noexcept
+    constexpr Parser_Error(std::int64_t xline, std::size_t xoffset, std::size_t xlength, Code xcode) noexcept
       : m_line(xline), m_offset(xoffset), m_length(xlength), m_code(xcode)
       {
       }
 
   public:
-    constexpr std::uint32_t line() const noexcept
+    constexpr std::int64_t line() const noexcept
       {
         return this->m_line;
       }
